+++ RELEASES.md
@@ -1,4 +1,4 @@
-# Release v0.4.3.3b 2025-01-14
+# Release v0.4.3.3b 2025-01-17
 
 ## New Features
 1. improve conversions (v0.4.3.3b 2025-01-05)
@@ -23,6 +23,8 @@
 9. rework replication (v0.4.3.3b 2025-01-13)
 10. performance improvements (v0.4.3.3b 2025-01-14)
 11. example improvements (v0.4.3.3b 2025-01-14)
+12. add local socket support, just start with unix socket (i.e. `/tmp/barch') on port 0 (v0.4.3.3b 2025-01-17)
+13. add luajit build support interfaces are suitable of for openresty and nginx (v0.4.3.3b 2025-01-17)
 
 ## Fixes
 1. fix conversions (v0.4.3.3b 2025-01-05)